assignment/Assignment-4/pyproject.toml
[project]
name = "ass4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pymongo>=4.15.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]
